+++ ChallengeIBGE.Core/Contexts/AddressContext/UseCases/UpdateAddress/Request.cs
@@ -2,4 +2,4 @@
 
 namespace ChallengeIBGE.Core.Contexts.AddressContext.UseCases.UpdateAddress;
 
-public record Request(Guid Id, string UpdatedCity, string UpdatedState, long UpdatedIbgeCode) : IRequest<Response>;
+public record Request(Guid Id, string UpdatedCity, string UpdatedState, int UpdatedIbgeCode) : IRequest<Response>;
